requirements.txt
beautifulsoup4==4.12.2
matplotlib==3.5.2
opencv_python==4.7.0.72
python-dotenv==1.0.0
requests==2.28.1
selenium==4.10.0
webdriver_manager==3.8.5
lxml==4.9.2
PyMySQL==1.0.3
cryptography==41.0.1